fix(cli): actually disable rebuild-orogen-models by default
It got disabled on the Roby side, but the flags still were true
by default. The option is broken, and not extremely useful so
far (i.e. fixing it is low-priority)

+++ lib/syskit/log/datastore.rb
@@ -119,10 +119,8 @@ module Syskit::Log
 
         # Remove an existing dataset
         def delete(digest)
+            cache_path_of(digest).rmtree if cache_path_of(digest).exist?
             core_path_of(digest).rmtree
-            if cache_path_of(digest).exist?
-                cache_path_of(digest).rmtree
-            end
         end
 
         # The full path to a dataset
